+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add React 16 support.
 * If at least intro.js 2.8.0 is used, `false` can be returned in `onBeforeChange` to prevent transitioning to the next / previous step.
 * Add the `onPreventChange` callback called when a transition is prevented by returning `false` in `onBeforeChange`.
+* Add the `onBeforeExit` callback to prevent exiting the intro.
 
 ## 0.1.5